TW-2268: refactor
