+++ src/store/partners-promotion/partners-promotion-reducers.ts
@@ -2,12 +2,7 @@ import { createReducer } from '@reduxjs/toolkit';
 
 import { AD_HIDING_TIMEOUT } from 'src/hooks/use-ad-temporary-hiding.hook';
 
-import {
-  hidePromotionAction,
-  setAdsEnabledEventSentAction,
-  setHasSeenRewardsAnnouncementAction,
-  togglePartnersPromotionAction
-} from './partners-promotion-actions';
+import { hidePromotionAction, togglePartnersPromotionAction } from './partners-promotion-actions';
 import { partnersPromotionInitialState } from './partners-promotion-state';
 
 export const partnersPromotionReducers = createReducer(partnersPromotionInitialState, builder => {
@@ -26,11 +21,5 @@ export const partnersPromotionReducers = createReducer(partnersPromotionInitialS
     }
 
     promotionHidingTimestamps[pathname] = timestamp;
-  });
-  builder.addCase(setHasSeenRewardsAnnouncementAction, state => {
-    state.hasSeenRewardsAnnouncement = true;
-  });
-  builder.addCase(setAdsEnabledEventSentAction, state => {
-    state.isAdsEnabledEventSent = true;
   });
 });
