+++ src/store/partners-promotion/partners-promotion-state.mock.ts
@@ -2,7 +2,5 @@ import type { PartnersPromotionState } from './partners-promotion-state';
 
 export const mockPartnersPromotionState: PartnersPromotionState = {
   isEnabled: false,
-  promotionHidingTimestamps: {},
-  hasSeenRewardsAnnouncement: true,
-  isAdsEnabledEventSent: true
+  promotionHidingTimestamps: {}
 };
