+++ src/store/partners-promotion/partners-promotion-state.ts
@@ -1,13 +1,9 @@
 export interface PartnersPromotionState {
   isEnabled: boolean;
   promotionHidingTimestamps: Record<string, number>;
-  hasSeenRewardsAnnouncement: boolean;
-  isAdsEnabledEventSent: boolean;
 }
 
 export const partnersPromotionInitialState: PartnersPromotionState = {
   isEnabled: true,
-  promotionHidingTimestamps: {},
-  hasSeenRewardsAnnouncement: false,
-  isAdsEnabledEventSent: false
+  promotionHidingTimestamps: {}
 };
