+++ src/store/rewards/rewards-selectors.ts
@@ -3,3 +3,6 @@ import { useSelector } from '../selector';
 export const useTkeyRewardsStatsSelector = () => useSelector(state => state.rewards.tkeyStats.data);
 
 export const useIsTkeyRewardsStatsLoadingSelector = () => useSelector(state => state.rewards.tkeyStats.isLoading);
+
+export const useHasSeenRewardsAnnouncementSelector = () =>
+  useSelector(state => state.rewards.hasSeenRewardsAnnouncement);
